requirements.txt
coverage==5.5
Faker==14.0.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.25.0
packaging==23.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.31.0
retrying==1.3.4
SQLAlchemy==2.0.2
Werkzeug==2.2.2
